requirements.txt
tensorflow==2.15.0.post1
numpy==1.26.4
scikit-learn==1.4.0
pillow==10.3.0
dask==2024.5.1
matplotlib==3.9.0
seaborn==0.13.1
mlflow==2.11.3
customtkinter==5.2.2
soundcard==0.4.3
soundfile==0.12.1
librosa==0.10.2.post1
keras_cv==0.8.1